concurrent: split _kernels.py into launched kernels + _tiles.py device primitives
Move the per-tile @triton.jit device functions (_grouped_coords, _gemm_tile,
_all_gather_tile, _all_reduce_tile, _reduce_scatter_tile, _all_to_all_tile,
_broadcast_tile, _comm_tile_ext) into a new iris/concurrent/_tiles.py; _kernels.py
now holds only the 7 launched work-stealing kernels and imports the tiles it calls.
Pure code organization -- no behavior change (emitted kernels are identical).

Validated: imports resolve across modules and 22/22 tests/ccl/test_concurrent_fused
(all 5 collectives x fused/concurrent).
